Changed makefile

+++ allreduce_large_buffer/allreduce.cpp
@@ -89,7 +89,7 @@ bool checkResults(T *ptr, size_t count, int rank, bool check_output, int repetit
     if (check_output)
     {
 #pragma omp parallel for
-        for (int i = 0; i < count; ++i) {
+        for (unsigned long i = 0; i < count; ++i) {
             T in[MAX_RANK];
             //create the input values
             for (int r = 0; r < world; ++r)
@@ -116,7 +116,7 @@ bool checkResults(T *ptr, size_t count, int rank, bool check_output, int repetit
             //check the kernel result against the reference result
             if (ptr[i] != (T)sum)
             {
-                if(returnval) printf("rank%d: mismatched at index %d, ref=%f, kernel=%f\n", rank, i, (double)sum, (double)ptr[i]);
+                if(returnval) printf("rank%d: mismatched at index %d, ref=%f, kernel=%f\n", rank, (int)i, (double)sum, (double)ptr[i]);
                 returnval = false;
             }
             else
@@ -124,7 +124,7 @@ bool checkResults(T *ptr, size_t count, int rank, bool check_output, int repetit
                 static bool printed = false;
                 if (!printed)
                 {
-                    printf("rank%d: Matched at index %d, ref=%f, kernel=%f\n", rank, i, (double)sum, (double)ptr[i]);
+                    printf("rank%d: Matched at index %d, ref=%f, kernel=%f\n", rank, (int)i, (double)sum, (double)ptr[i]);
                     printed = true;
                 }
             }
