+++ allreduce_large_buffer/allreduce.h
@@ -117,7 +117,7 @@ void load_input_to_temp_buffer(int idx, void* inout_buffer, uint32_t size, int t
     if (read_offset + SIMD_COMPUTE * TEMP_WORLD > size)
     {
         int count = (size - read_offset + SIMD_COMPUTE - 1) / SIMD_COMPUTE;
-        for (uint32_t i = 0; i < count; i++)
+        for (int i = 0; i < count; i++)
         {
             buffer.template select<SIMD_COMPUTE, 1>(SIMD_COMPUTE * i) = lsc_block_load<T, SIMD_COMPUTE, lsc_data_size::default_size, cache_hint::uncached, cache_hint::uncached>
                 ((T *)inout_buffer + read_offset + i * SIMD_COMPUTE);
@@ -127,7 +127,7 @@ void load_input_to_temp_buffer(int idx, void* inout_buffer, uint32_t size, int t
     else
     {
 #pragma unroll
-        for (uint32_t i = 0; i < TEMP_WORLD; i++)
+        for (unsigned int i = 0; i < TEMP_WORLD; i++)
         {
             buffer.template select<SIMD_COMPUTE, 1>(SIMD_COMPUTE * i) = lsc_block_load<T, SIMD_COMPUTE, lsc_data_size::default_size, cache_hint::uncached, cache_hint::uncached>
                 ((T *)inout_buffer + read_offset + i * SIMD_COMPUTE);
@@ -315,7 +315,7 @@ void write_output(int idx, void* inout_buffer, uint32_t size, int threads_alread
     if (write_offset + SIMD_COMPUTE * TEMP_WORLD > size)
     {
         int count = (size - write_offset + SIMD_COMPUTE - 1) / SIMD_COMPUTE;
-        for (uint32_t i = 0; i < count; i++)
+        for (int i = 0; i < count; i++)
         {
             lsc_block_store<T, SIMD_COMPUTE, lsc_data_size::default_size, cache_hint::uncached, cache_hint::uncached>
                 (write_ptr + i * SIMD_COMPUTE, results.template select<SIMD_COMPUTE, 1>(SIMD_COMPUTE * i));
@@ -484,7 +484,7 @@ public:
                 if (iter == 1) //if second iteration, then handle the partial usage of the temp buffer
                 {
                     //if there is little more left to compute, then finish them
-                    if (outerloop_iter_count * max_elements_per_MAX_COUNT < size)
+                    if (outerloop_iter_count * max_elements_per_MAX_COUNT < (int)size)
                     {
                         start = outerloop_iter_count;
                         end = start + 1;
@@ -558,7 +558,7 @@ public:
                                         for (int inner_iter = 0; inner_iter < innerloop_iter_count; inner_iter++)
                                         {
                                             int index = idx + inner_iter * HW_THREAD_COUNT;
-                                            if (index >= total_threads_needed)
+                                            if ((uint32_t)index >= total_threads_needed)
                                                 break;
 
                                             switch (temp_world)
@@ -588,7 +588,7 @@ public:
                                         for (int inner_iter = 0; inner_iter < innerloop_iter_count; inner_iter++)
                                         {
                                             int index = idx + inner_iter * HW_THREAD_COUNT;
-                                            if (index >= total_threads_needed)
+                                            if ((uint32_t)index >= total_threads_needed)
                                                 break;
 
                                             switch (temp_world)
@@ -618,7 +618,7 @@ public:
                                         for (int inner_iter = 0; inner_iter < innerloop_iter_count; inner_iter++)
                                         {
                                             int index = idx + inner_iter * HW_THREAD_COUNT;
-                                            if (index >= total_threads_needed)
+                                            if ((uint32_t)index >= total_threads_needed)
                                                 break;
 
                                             switch (temp_world)
@@ -648,7 +648,7 @@ public:
                                         for (int inner_iter = 0; inner_iter < innerloop_iter_count; inner_iter++)
                                         {
                                             int index = idx + inner_iter * HW_THREAD_COUNT;
-                                            if (index >= total_threads_needed)
+                                            if ((uint32_t)index >= total_threads_needed)
                                                 break;
 
                                             switch (temp_world)
@@ -678,7 +678,7 @@ public:
                                         for (int inner_iter = 0; inner_iter < innerloop_iter_count; inner_iter++)
                                         {
                                             int index = idx + inner_iter * HW_THREAD_COUNT;
-                                            if (index >= total_threads_needed)
+                                            if ((uint32_t)index >= total_threads_needed)
                                                 break;
 
                                             switch (temp_world)
@@ -831,7 +831,7 @@ private:
             &send_buf, sizeof(send_buf), MPI_BYTE, recv_buf, sizeof(send_buf), MPI_BYTE, MPI_COMM_WORLD);
 
         
-        for (uint32_t i = 0; i < world; i++)
+        for (int i = 0; i < world; i++)
         {
             // Step 4: Prepare pid file descriptor of next process
             auto* peer = recv_buf + i;
